[dartdevc] Fix printing of generic function types
The toString() method of generic function runtime type objects should
not print type argument bounds when they are a top type.

In weak mode the bounds will also be elided when they are Object or
Object*.

Updated the DDC default (no type bounds emitted from the compiler) to
represent Object* instead of dynamic. This ensures that we properly
assign the bounds to the type of functions that have been declared
in legacy libraries. All other top type bounds are emitted explicitly
from the compiler.

Fixes: #40938
Change-Id: I3b35e974d79e2bbca75b54079e42f69cbaf2dfa2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/139040
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>