Improve the relevance scores for instance type members
The "super matches" feature should have been disabled (-1.0) when the
target wasn't `super`, but it wasn't. That artificially lowered the
score for methods. But it wasn't being used (even when it should have
been) for getters and setters, so their scores weren't being lowered.
That accounted for the poor mrr I was seeing. This fixes both problems
so that the feature is now used everywhere when it should be and not
when it shouldn't be.

Change-Id: I541a9b1ed524e8b2ed2c0c7f39e7bca786e12ea9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/139380
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Reviewed-by: Jaime Wren <jwren@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>