[nnbd_migration] Split apart running tasks from sending dartfix response
This enables us to get code that we can execute to rerun dartfix in
later CLs.

One note is that this won't pick up on certain file changes, etc. I
think a better long-term solution is to rerun the entire DartFix code on
rerun, and that dartfix response can be returned directly to the js
front end. Its possible that the types of file changes this will miss
are extremely rare, such as changed .packages files only etc.

This would require a bit more untangling, things like managing the way that
dartfix will start a server each time, and ensuring that we get a
DartFix respnse *and* a migration state, so this CL seems worth landing
in the meantime. Filed as #41017

Change-Id: If629cac41149728b34f6eb9785a7cea543a23a6f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/139306
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Janice Collins <jcollins@google.com>