[dart2js] remove invalid assertion in dart2js batch runner.
Turns out the invariant is not respected in practice. This is what caused hundreds of flaky failures in the old-rti bot when I added this assert.

The batchArguments include experiments whenever the first test in a shard has an experiment flag. If that happens and most other test in the same shard don't have the experiment, all those tests will appear to crash with this assertion.  Because the assertion is outside the test itself, we get almost no feedback on the test logs.

TBR=fishythefish@google.com

Change-Id: I1230d79ed3cccbe388dddf19b5e8aa18c50f7329
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/139566
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>