[CFE] Add test that combine different scenarios
This CL adds a test that combine different scenarios for compilation,
and check that we

* don't crash
* verify
* can serialize

The current combinations allow for combining an "outer level"
(top level, class level, extension level) with an "inner level"
(methods, fields, either static or not) with an expression.

This yields a large number of tests that has already shown some errors
(that was corrected before uploading this CL though).

The test can certainly be update later to uncover more territory.

Change-Id: Ic4b46ff3e26255c88c8eb8d8ce45f0257ffbc993
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/139621
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>