Use context type in the static member contributor
I couldn't think of any other features related to static members. Please
let me know if you can think of any.

I need to modify the metrics utility to be able to differentiate between
static and instance member references in order to measure the impact of
this approach. I'll look into doing that today. I suspect that we'll
eventually need it to split out results in many more ways as well.

Also updated some comments in an attempt to make it easier in the future
to determine which contributor to look at for bug fixes or enhancements.

Change-Id: Ieef62c6189bbfbd35283abdf35b8cbdd17746ee8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/139820
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>