Add dart2js test directories to the CFE opt-out whitelist.
This doesn't reduce the technical debt of having the whitelist in the
first place, but it correctly treats those tests as being opted out.

This doesn't address "compiler/dartdevc_native". That one's a little
harder because it contains a mixture of NNBD and legacy libraries. We
might want to put "@dart=2.7" tags in there for the legacy ones. I'll
look into that in a separate CL.

But this should at least get the CFE handling some more of the tests
correctly.

Change-Id: I4db0d671ac097eba27bfbeedb296a8526043622a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/139868
Reviewed-by: Bob Nystrom <rnystrom@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Auto-Submit: Bob Nystrom <rnystrom@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>