[ddc] Disable type check failure suggestions
These messages are created when a type check of a generic class fails
and were designed to help in the Dart 2.0 migration by suggesting a type
that would work. They should either be updated to suggest types
correctly in the presence of the Null Safety features, or be removed
entirely if they are no longer helpful.

Issue: https://github.com/dart-lang/sdk/issues/40326
Change-Id: Ic8ecc83cad534471a88a16baf22035d90b0336ca
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/140341
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>