Revert "Use the 'any' constraint for the recommended dep on package:test."
This reverts commit c27af755043b0b82f52e18c24164893cdd9ac24c.

Reason for revert: There's not consensus on the `^1.0.0` vs. `any` constraint on package:test; see https://github.com/dart-lang/stagehand/pull/657 for the conversation. I'm reverting this change for now so that stagehand and dartdev are in agreement. We can update them both if we decide that `any` is a better constraint to recommand.

Original change's description:
> Use the 'any' constraint for the recommended dep on package:test.
> 
> Change-Id: I4052cf0653d827795e120b9b2beecc285935973a
> Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/139680
> Reviewed-by: Gary Roumanis <grouma@google.com>
> Reviewed-by: Jaime Wren <jwren@google.com>
> Reviewed-by: Michael Thomsen <mit@google.com>
> Commit-Queue: Devon Carew <devoncarew@google.com>

TBR=mit@google.com,devoncarew@google.com,jwren@google.com,grouma@google.com

# Not skipping CQ checks because original CL landed > 1 day ago.

Change-Id: I60c09b399530f4860f59bd31535d29ca3adb70f3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/140764
Reviewed-by: Jaime Wren <jwren@google.com>
Commit-Queue: Devon Carew <devoncarew@google.com>