[vm/compiler] Force members with @pragma('vm:entry-point') annotation to have boxed calling convention
So far the unboxing support in TFA disabled unboxing for natives and
relied on TFA inferring entrypoint parameters to be nullable (and
therefore boxed).

Though the added support for unboxing return values missed to explicitly
disable unboxing for return values of entry points.
  => This CL fixes that.

This is also a preparation for NNBD when parameters to entry points are not
necessarily nullable anymore.

Change-Id: Ia80a6acc8e803732b81665ebab14a9bfa720be22
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/140866
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Daco Harkes <dacoharkes@google.com>
Reviewed-by: Alexander Markov <alexmarkov@google.com>