[infra] Put VM canary and product builders conditionally on the CQ
They should still be added on all VM, CFE and test CLs, but will no
longer be added to unrelated CLs (e.g. analyzer changes). Both of them
are Linux builders, and the product builder also runs co19 tests. The
release builder is still on the CQ and should provide sufficient
coverage in case the patterns miss some changes that affect the VM.

Change-Id: I0a05fefd76a0659fae209af144163d0fe7ea10d1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/141124
Reviewed-by: William Hesse <whesse@google.com>