[cfe] Use legacy erasure before argument inference
If calling an opt-in method from opt-out we would use the opt-in
parameter types as type context for the arguments leading to
opt-in inference in opt-out libraries. With this change the
callee type is legacy erased before inference when the invocation
is in an opt-out library.

Closes #41180

Change-Id: I55a0a4f055f500652c154c020a175fe12892abb9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/140861
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>