[nnbd_migration] Save line number of details to preserve on reload.
When clicking on a region, store its offset & line in the URL via
`pushState`, so that on reload (after rerun migration or after apply
edit, or just in general) the context isn't lost.

Note that offset is not enough. Basically, every navigable element gets
a class `o$offset`, and if the page is loaded with some offset such
that no element with class `o$offset` exists, it falls back to line
number. And if there were edits mode before the region since reload,
there will only be an element `o$offset` if you are lucky.

Change-Id: Ic5fd83aceb4e8731348e70c039fec2ca2a6ab9e4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/141465
Commit-Queue: Mike Fairhurst <mfairhurst@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>