Adjust member signature expectations for mixed inheritance
The rule NNBD_TOP_MERGE(S*,T) = NNBD_TOP_MERGE(S,T) and the converse
imply that when a legacy member signature and a null-safe member
signature are brought together by two different superinterfaces,
the null-safe signature prevails (so the legacy signature is ignored).

This CL makes adjustments to 'legacy_resolves_conflict_3_test.dart'
to match this. Note that this means that a class like DiB3 has a
member signature `int Function(int) get a` even though it inherits an
implementation with signature `int? Function(int?) get a`, so we may
have to make such classes an error. [Edit: Based on the notion of
a 'class implementation', which is consolidated in the same sense as
the class interface, but which records the concrete member
declarations available in superclasses, with the declaration in the
nearest superclass prevailing over declarations in more remote ones,
the conflict disappears in the usual way: The class implementation
of a legacy class is legacy-erased. So there are no errors for those
classes after all.]

Change-Id: Ie72e6b80064764fa722304f32bee5e9188710a2c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/142372
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Leaf Petersen <leafp@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Erik Ernst <eernst@google.com>